Move inner classes out and split out a bit.
