+++ std-bits/tableau/src/main/java/org/enso/tableau/HyperTable.java
@@ -1,0 +1,5 @@
+package org.enso.tableau;
+
+// ** Record type for representing a Hyper table. */
+public record HyperTable(String schema, String name) {
+}
